moved slider higher to switch between 2024/2025 supporter/sponsor info
