+++ app/(pages)/starlight/page.tsx
@@ -1,12 +1,6 @@
 import React from "react";
 import PageSection from "../../../components/PageSection";
 import { Box, Grid, StepIndicator, stepIndicatorClasses } from "@mui/joy";
-import {
-  SponsorInfo,
-  starlightProjectData,
-  starlightSupporterData,
-} from "../../../data";
-import { ProjectCard } from "../../../components/ProjectCard";
 import type { Metadata } from "next";
 import {AspectRatio, Link, Stack, Typography} from "@mui/material";
 import { StarlightProjectsSection } from "../../../components/StarlightProjectsSection";
@@ -16,7 +10,6 @@ import HailRoundedIcon from "@mui/icons-material/HailRounded";
 import KeyboardVoiceRoundedIcon from "@mui/icons-material/KeyboardVoiceRounded";
 import WavingHandRoundedIcon from "@mui/icons-material/WavingHandRounded";
 import EmojiEventsIcon from "@mui/icons-material/EmojiEvents";
-import Image from "next/image";
 
 export const metadata: Metadata = {
   title: "Starlight | DevSoc UNSW",
@@ -111,84 +104,9 @@ export default function StarlightProjectsPage() {
                 </Typography>
                 <StarlightStepper />
                 <Box sx={{ pb: 4}} />
-                <Typography level="h3" py={2} textAlign={"center"}>
-                    Supporters
-                </Typography>
-                <Stack
-                    display="grid"
-                    gridTemplateColumns={{
-                        sm: "repeat(2, 1fr)",
-                        md: "repeat(4, 1fr)",
-                    }}
-                    marginBottom={5}
-                    sx={{ gridGap: "20px" }}
-                >
-                    {starlightSupporterData["2024"].supporterLogos.map((sponsor, idx) => {
-                        return renderLogoRows(idx, sponsor, "small");
-                    })}
-                </Stack>
-                <Typography level="h3" py={2} textAlign={"center"}>
-                    2025 Industry Guests
-                </Typography>
-                <Stack
-                    display="grid"
-                    gridTemplateColumns={{
-                        xs: "repeat(auto-fit, minmax(200px, 1fr))",
-                        md: "repeat(auto-fit, minmax(200px, 1fr))",
-                        xl: "repeat(auto-fit, minmax(200px, 1fr))",
-                    }}
-                    marginBottom={8}
-                    sx={{ gridGap: "32px" }}
-                >
-                    {(starlightSupporterData["2025"].industryLogos || [])
-                        .map((sponsor, idx) => renderLogoRows(idx, sponsor))}
-                </Stack>
-                <Typography level="h3" py={2} textAlign={"center"}>
-                    2025 Society Partners
-                </Typography>
-                <Stack
-                    display="grid"
-                    gridTemplateColumns={{
-                        sm: "repeat(2, 1fr)",
-                        md: "repeat(4, 1fr)",
-                    }}
-                    marginBottom={8}
-                    sx={{ gridGap: "32px" }}
-                >
-                    {(starlightSupporterData["2025"].societyLogos || [])
-                        .map((sponsor, idx) => renderLogoRows(idx, sponsor, "small"))}
-                </Stack>
             </PageSection>
             
             <StarlightProjectsSection />
         </>
     );
-}
-
-function renderLogoRows(idx: number, sponsor: SponsorInfo, size: "small" | "large" = "large") {
-  const isRamsoc = sponsor.name && sponsor.name.toLowerCase().includes("ramsoc");
-  const sizeProps = isRamsoc
-    ? { height: 40, maxWidth: 100 }
-    : size === "small"
-      ? { height: 50, maxWidth: 130 }
-      : { height: 50, maxWidth: 170 };
-  return (
-    <AspectRatio
-      key={idx}
-      variant="plain"
-      ratio="6/3"
-      objectFit="contain"
-      sx={{
-        display: "flex",
-        margin: "auto",
-        width: "100%",
-        padding: 0.5,
-        ...sizeProps,
-      }}
-    >
-      <Link target="_blank" href={sponsor.url}>
-        <Image src={sponsor.logo} alt={sponsor.name} fill priority />
-      </Link>
-    </AspectRatio>
-  );
 }
